+++ tests/dwd/observations/test_metaindex.py
@@ -3,14 +3,12 @@ import pytest
 import requests
 
 from wetterdienst.dwd.metadata.column_names import DWDMetaColumns
-from wetterdienst.dwd.observations import (
-    DWDObservationParameterSet,
-    DWDObservationPeriod,
-    DWDObservationResolution,
-)
+from wetterdienst.dwd.observations import DWDObservationParameterSet
 from wetterdienst.dwd.observations.metaindex import (
     create_meta_index_for_climate_observations,
 )
+from wetterdienst.metadata.period import Period
+from wetterdienst.metadata.resolution import Resolution
 
 
 @pytest.mark.remote
@@ -19,8 +17,8 @@ def test_meta_index_creation():
     # Existing combination of parameters
     meta_index = create_meta_index_for_climate_observations(
         DWDObservationParameterSet.CLIMATE_SUMMARY,
-        DWDObservationResolution.DAILY,
-        DWDObservationPeriod.HISTORICAL,
+        Resolution.DAILY,
+        Period.HISTORICAL,
     )
 
     assert not meta_index.empty
@@ -29,8 +27,8 @@ def test_meta_index_creation():
     with pytest.raises(requests.exceptions.HTTPError):
         create_meta_index_for_climate_observations(
             DWDObservationParameterSet.CLIMATE_SUMMARY,
-            DWDObservationResolution.MINUTE_1,
-            DWDObservationPeriod.HISTORICAL,
+            Resolution.MINUTE_1,
+            Period.HISTORICAL,
         )
 
 
@@ -39,8 +37,8 @@ def test_meta_index_1mph_creation():
 
     meta_index_1mph = create_meta_index_for_climate_observations(
         DWDObservationParameterSet.PRECIPITATION,
-        DWDObservationResolution.MINUTE_1,
-        DWDObservationPeriod.HISTORICAL,
+        Resolution.MINUTE_1,
+        Period.HISTORICAL,
     )
 
     assert meta_index_1mph.loc[