+++ wetterdienst/dwd/radar/metadata/resolution.py
@@ -1,12 +1,13 @@
+# -*- coding: utf-8 -*-
+# Copyright (c) 2018-2020 earthobservations
+# Copyright (c) 2018-2020 Andreas Motl <andreas.motl@panodata.org>
+# Copyright (c) 2018-2020 Benjamin Gutzmann <gutzemann@gmail.com>
 from enum import Enum
+
+from wetterdienst.metadata.resolution import Resolution
 
 
 class DWDRadarResolution(Enum):
-    """
-    enumeration for granularity/resolution of the weather
-    observations stored on dwd server
-    """
-
-    MINUTE_5 = "5_minutes"
-    HOURLY = "hourly"
-    DAILY = "daily"
+    MINUTE_5 = Resolution.MINUTE_5.value
+    HOURLY = Resolution.HOURLY.value
+    DAILY = Resolution.DAILY.value